Updated To Remove "Victory" and "Failure" Components
Victory and failure components may be added later on if the original website receives an update to track the number of incorrect actions a player has taken.

As of time manual tracking has to be done through the use of timpepoints offset from the original by a microsecond. However, to do this there must be a punishment time offset from the original punishment time that points to the time offset from the original. This has to be done N number of times where N = Health. 10 timepoints with 3 options and 3 health results in 104 choices. This grows without bound as timepoints, options, and health increases.