Table View of Added Templates & Bug Fixes
Fixed centering issue where the buttons were placed directly next to each other.

Fixed issue in VideoEditors where a conditional video would point directly from a continue point to a return point instead of a microsecond before the return point. (This caused the video to not repeat actions after the used selected incorrectly and would just proceed to the next action).

Disabled adding templates with the same name twice

Added a table view of already added templates that updates as templates are added.

Fixed issue where webpage would do nothing if there is a error with the API.

Added new punishment module examples to the README